move pong logic to pong service
