+++ api/src/modules/game/game.module.ts
@@ -7,11 +7,12 @@ import { GameController } from './game.controller';
 import { GameGateway } from './game.gateway';
 import { GameService } from './game.service';
 import { NotifyModule } from '../notify';
+import { PongService } from './pong.service';
 
 @Module({
   imports: [UserModule, NotifyModule],
   controllers: [GameController],
-  providers: [GameGateway, GameService, PrismaService],
+  providers: [GameGateway, GameService, PrismaService, PongService],
   exports: [GameService],
 })
 export class GameModule {}
